Favorite quotes done :zap:
